Stdlib decoder isn't particularly fast, so switched to a different one with better performance. Some numbers:

```
$ benchstat /tmp/small_old.txt /tmp/small_new.txt
goos: darwin
goarch: arm64
pkg: github.com/vespa-engine/vespa/client/go/internal/vespa/document
                        │ /tmp/small_old.txt │         /tmp/small_new.txt          │
                        │       sec/op       │   sec/op     vs base                │
DocumentDecoderSmall-10          827.7n ± 1%   475.5n ± 0%  -42.55% (p=0.000 n=10)

$ benchstat /tmp/large_old.txt /tmp/large_new.txt
goos: darwin
goarch: arm64
pkg: github.com/vespa-engine/vespa/client/go/internal/vespa/document
                        │ /tmp/large_old.txt │          /tmp/large_new.txt          │
                        │       sec/op       │    sec/op     vs base                │
DocumentDecoderLarge-10          48.79µ ± 1%   16.76µ ± 13%  -65.65% (p=0.000 n=10)
```

@jonmv